requirements.txt
google-generativeai>=0.8.4